# sandboxd-orch/docs/orchestrator_swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Orchestrator API Server",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/externals": {
            "post": {
                "description": "Sets a single external endpoint for one node (1:1 per node).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orchestrator-node"
                ],
                "summary": "Create/Update external object",
                "parameters": [
                    {
                        "description": "External object request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateExternalObjectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.HealthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/nodes": {
            "get": {
                "description": "Lists all registered nodes with state and latest resource data.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orchestrator-node"
                ],
                "summary": "List nodes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.NodesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates or updates a node object from id/spec(ip,port).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orchestrator-node"
                ],
                "summary": "Create/Update node object",
                "parameters": [
                    {
                        "description": "Node object request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateNodeObjectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.NodeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/nodes/{id}": {
            "get": {
                "description": "Returns a single node by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orchestrator-node"
                ],
                "summary": "Get node",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.NodeResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes node registration and detaches related sandbox scheduling metadata.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orchestrator-node"
                ],
                "summary": "Delete node",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Force delete without node API calls (use only when node is already gone)",
                        "name": "force",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.DeleteNodeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/nodes/{id}/heartbeat": {
            "post": {
                "description": "Probes sandboxd health and node resources immediately.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orchestrator-node"
                ],
                "summary": "Trigger node heartbeat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.HeartbeatResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/nodes/{id}/reconcile": {
            "post": {
                "description": "Proxies manual reconcile trigger to selected node sandboxd.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orchestrator-proxy"
                ],
                "summary": "Proxy reconcile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ProxyResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/nodes/{id}/sandboxes": {
            "get": {
                "description": "Proxies sandbox list request to selected node sandboxd.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orchestrator-proxy"
                ],
                "summary": "Proxy list sandboxes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pagination cursor",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ProxyResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Proxies create-sandbox request directly to selected node sandboxd.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orchestrator-proxy"
                ],
                "summary": "Proxy create sandbox",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Sandbox create request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateSandboxRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ProxyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.DeleteNodeResponse": {
            "type": "object",
            "properties": {
                "deleted": {
                    "type": "string"
                }
            }
        },
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "handlers.HealthResponse": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean"
                }
            }
        },
        "handlers.HeartbeatResponse": {
            "type": "object",
            "properties": {
                "heartbeat": {
                    "type": "string"
                },
                "heartbeat_error": {
                    "type": "string"
                },
                "node": {
                    "$ref": "#/definitions/types.Node"
                },
                "resources": {},
                "status_error": {
                    "type": "string"
                }
            }
        },
        "handlers.NodeResponse": {
            "type": "object",
            "properties": {
                "node": {
                    "$ref": "#/definitions/types.Node"
                }
            }
        },
        "handlers.NodesResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Node"
                    }
                }
            }
        },
        "handlers.ProxyResponse": {
            "type": "object",
            "additionalProperties": {}
        },
...
        "model.CreateSandboxRequest": {
            "type": "object",
            "properties": {
                "containers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CreateContainerRequest"
                    }
                },
                "egress": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "ports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.PortMapping"
                    }
                },
                "readinessProbe": {
                    "$ref": "#/definitions/model.ReadinessProbeSpec"
                }
            }
        },
...
        "types.CreateExternalObjectRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "spec": {
                    "$ref": "#/definitions/types.ExternalObjectSpec"
                }
            }
        },
        "types.CreateNodeObjectRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "spec": {
                    "$ref": "#/definitions/types.NodeObjectSpec"
                }
            }
        },
...
        "types.ExternalObjectSpec": {
            "type": "object",
            "properties": {
                "external": {
                    "type": "string"
                },
                "node_id": {
                    "type": "string"
                }
            }
        },
...
    }
}